+++ src/main/java/chess/chessboard/Rank.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess.chessboard;
 
 import java.util.*;
 import java.util.stream.Collectors;
