+++ src/main/java/chess/chessboard/Side.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess.chessboard;
 
 public enum Side {
     WHITE,
