+++ src/test/java/chess/chessboard/ChessBoardTest.java
@@ -1,5 +1,6 @@
-package chess.domain;
+package chess.chessboard;
 
+import chess.piece.*;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
