+++ src/test/java/chess/chessboard/RankTest.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess.chessboard;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
